Intermediate Model Transformations: Fix ATL NPE in intermediate data model refinement
The NPE occurred when
  - the return type of a domain operation was mapped to a technology-specific type and
  - the original type was a data structure and
  - the data structure comprised aspects.

If these circumstances were met, the ATL VM threw an NPE in the postApply action and
for the refSetValue() call removed by this commit. The rationale seems to be that for
some reason the ATL transformation produced an intermediate aspect instance, in which
all fields were null. Without the call to refSetValue() in the
*IntermediateImportedTechnologySpecificType* rule, which directly has nothing to do
with intermediate aspects, however, the transformation seems to work as expected.