Releng: Don't clone LEMMA repository twice in build script.
In the build script, the LEMMA repository is first cloned by Jenkins, before we clone
it again to eventually run the build. However, this approach does not work for builds
triggered by PRs. That is, because our clone task is not aware of the PR and the build
will fail, because the branch "PR-x" does not exist in the cloned repository.

The first clone task executed by Jenkins does however consider PRs correctly and applies
their changes to the cloned tree, before it triggers the build. Therefore, this commit
removes our second clone task from the build script, so that we completely rely on Jenkins'
clone task to correctly handle PR-induced builds.