Java Base Generator: Replace when() in overridden equals() implementations with if-conditionals.
The rationale for this change is the fact that when() invokes equals() on conditional branches.
Consequently, overridden equals() implementations may be invoked again, which can result in
in wrong results (e.g., an overridden equals(other) returns true even if other is effectively
not equal to the current object) or, even worse, lead to recursions of equals() calls.