Live Validation Framework: Refactor towards loosely coupled Java implementation
We split the current implementation of the Live Validation Framework into the
following modules:
    - util: Shared utilities.
    - model: Shared classes for Live Validation issue representation.
    - protocol: Extensions to LSP specific to Live Validation.
    - client: Framework components for Live Validation client implementation.
    - server: Standalone executable implementation of Live Validation server.

The refactoring facilitates the usage of the Live Validation Framework by both
the Eclipse editor plugin (Xtend-based) and LEMMA's Model Processing Framework
(Kotlin-based). In particular, we split the server-related from the
client-related parts and do no longer require the installation of
Kotlin-specific artifacts via a dedicated Eclipse updatesite.